Downloading monero-rpc-wallet during the build process. Fixing paths.
During runtime the paths were broken in an executable form because it
is expected to have files in the _internal folder.

At the moment it seems as though the create_wallet call to the wallet
rpc server isn't working for some reason. Difficult to debug since
there's not any indication as to what the actual problem is.