Saving some progress trying to implement pystray, making views snappy.
Having some troubles with pystray on my machine at least. It may be a
KDE/xorg specific issue, but it seems that I'm not the only one.
https://github.com/moses-palmer/pystray/issues/175 The implementation
is commented out for now and there's some comments about some of my
research.

Refactored the view code a bit to be snappier while switching between
views. Previously we would build/destroy the view every time we would
switch between them which would take a while and cause noticable lag.
Now we are creating all of the views when we start the application and
switching between them by making all of the elements invisible with
`grid_remove` and making them visible again with `grid`. This works
really well. There does seem to be some slight differences in styling
from before, but they're minor and solveable.