Fixing the icon issue by using pyqt5, adding some features to history.
Pystray doesn't work properly on KDE Linux. Qt should be much more
reliable for cross platform development.

The history screen wasn't showing the proper currencies, nor was it
being updated on a regular basis so when a new payment would come in,
you wouldn't see it until restarting the application, that has been
fixed and now it updates every 5 seconds.
It also starts to try to present the correct currency symbol, though
I'll need to double check that it does currency conversion properly.

This also adds an activated flag to views so we can do things on the
initial build of a view, and different things every time the view
is switched back to via reactivate.