Adding observer for rpc server readiness to get transactions.
When loading the application and navigating to the history page, it
would take 5 seconds from rendering the page to get the list of txs.
This will do a first fetch of the transactions as soon as possible
by making the request as soon as the wallet rpc has started.

This will also correctly convert the amounts of previous transactions.