Adding test for the pay flow for a payment request.
This also adds adding of subscriptions to the subscription view.
Previously this was done at the build time of the view as that only
happened when a view was navigated to. Since views are long lived now
we need to make sure it gets updated.

This also adds a testing config and mocking for it so we can use a
different config file than the one that is being used locally. There
are a lot of mocks in different files for different tests now and it
is quickly becoming difficult to manage. Ideally we can figure out
some common mocks and easier ways of mocking out tests without having
to declare all of them for ever test.